change ui to get path and restructure project
