+++ src/main/java/com/solvd/navigator/service/RoadService.java
@@ -1,4 +1,4 @@
-package com.solvd.navigator.services;
+package com.solvd.navigator.service;
 
 import com.solvd.navigator.dao.IRoadDao;
 import com.solvd.navigator.dao.jdbc.RoadDao;
