+++ src/main/java/com/solvd/navigator/util/AllPairShortestPath.java
@@ -1,9 +1,8 @@
-package com.solvd.navigator.model.util;
+package com.solvd.navigator.util;
 
 import com.solvd.navigator.model.PathContainer;
-import com.solvd.navigator.services.StationService;
+import com.solvd.navigator.service.StationService;
 
-import java.io.*;
 import java.lang.*;
 import java.util.*;
 
@@ -34,7 +33,8 @@ public class AllPairShortestPath {
     }
 
     public PathContainer floydWarshallWithPath(int[][] dist, int startStation, int endStation) {
-
+        startStation -= 1;
+        endStation -=1;
         //Path matrix
         int[][] pathMatrix = new int[V][V];
 
@@ -64,7 +64,7 @@ public class AllPairShortestPath {
         return new PathContainer(dist[startStation][endStation], pathFromAtoB);
     }
 
-    private List<Integer> pathFromAtoB(int[][] paths, int startStation, int endStation) {
+    public List<Integer> pathFromAtoB(int[][] paths, int startStation, int endStation) {
 
         //Initialize list to add stations to follow to arrive at the final station
         List<Integer> list = new LinkedList<>();
